+++ noodles-gff/examples/gff_write.rs
@@ -8,7 +8,7 @@ use noodles_gff::{self as gff, Line};
 
 fn main() -> io::Result<()> {
     let stdout = io::stdout().lock();
-    let mut writer = gff::Writer::new(stdout);
+    let mut writer = gff::io::Writer::new(stdout);
 
     let version = gff::Directive::GffVersion(Default::default());
     writer.write_directive(&version)?;