+++ noodles-gff/src/io.rs
@@ -1,5 +1,6 @@
 //! GFF I/O.
 
 pub(crate) mod reader;
+mod writer;
 
-pub use self::reader::Reader;
+pub use self::{reader::Reader, writer::Writer};